Only update answer collection once all pending assemblies are complete
Suggested fix for SOL-3139. When assembling templates that contain a variable with "Save in answer file" turned off: if the parent template also has an assemble instruction, the answers that would not usually be saved in an answer file must be passed to the child template when it is assembled. The fix changes when the answer collection is updated so that answers are retained until after the last child template is assembled.